+++ cloud-controller-manager/do/loadbalancers.go
@@ -136,6 +136,10 @@ const (
 	// be enabled. Defaults to false.
 	annDOEnableProxyProtocol = "service.beta.kubernetes.io/do-loadbalancer-enable-proxy-protocol"
 
+	// annDOEnableBackendKeepalive is the annotation specifying whether HTTP keepalive connections
+	// should be enabled to backend target droplets. Defaults to false.
+	annDOEnableBackendKeepalive = "service.beta.kubernetes.io/do-loadbalancer-enable-backend-keepalive"
+
 	// defaultActiveTimeout is the number of seconds to wait for a load balancer to
 	// reach the active state.
 	defaultActiveTimeout = 90
@@ -625,23 +629,29 @@ func (l *loadBalancers) buildLoadBalancerRequest(ctx context.Context, service *v
 		return nil, err
 	}
 
+	enableBackendKeepalive, err := getEnableBackendKeepalive(service)
+	if err != nil {
+		return nil, err
+	}
+
 	var tags []string
 	if l.resources.clusterID != "" {
 		tags = []string{buildK8sTag(l.resources.clusterID)}
 	}
 
 	return &godo.LoadBalancerRequest{
-		Name:                lbName,
-		DropletIDs:          dropletIDs,
-		Region:              l.region,
-		ForwardingRules:     forwardingRules,
-		HealthCheck:         healthCheck,
-		StickySessions:      stickySessions,
-		Tags:                tags,
-		Algorithm:           algorithm,
-		RedirectHttpToHttps: redirectHTTPToHTTPS,
-		EnableProxyProtocol: enableProxyProtocol,
-		VPCUUID:             l.resources.clusterVPCID,
+		Name:                   lbName,
+		DropletIDs:             dropletIDs,
+		Region:                 l.region,
+		ForwardingRules:        forwardingRules,
+		HealthCheck:            healthCheck,
+		StickySessions:         stickySessions,
+		Tags:                   tags,
+		Algorithm:              algorithm,
+		RedirectHttpToHttps:    redirectHTTPToHTTPS,
+		EnableProxyProtocol:    enableProxyProtocol,
+		EnableBackendKeepalive: enableBackendKeepalive,
+		VPCUUID:                l.resources.clusterVPCID,
 	}, nil
 }
 
@@ -1090,6 +1100,22 @@ func getEnableProxyProtocol(service *v1.Service) (bool, error) {
 	}
 
 	return enableProxyProtocol, nil
+}
+
+// getEnableBackendKeepalive returns whether HTTP keepalive to target droplets should be enabled.
+// False is returned if not specified.
+func getEnableBackendKeepalive(service *v1.Service) (bool, error) {
+	enableBackendKeepaliveStr, ok := service.Annotations[annDOEnableBackendKeepalive]
+	if !ok {
+		return false, nil
+	}
+
+	enableBackendKeepalive, err := strconv.ParseBool(enableBackendKeepaliveStr)
+	if err != nil {
+		return false, fmt.Errorf("failed to parse backend keepalive flag %q from annotation %q: %s", enableBackendKeepaliveStr, annDOEnableBackendKeepalive, err)
+	}
+
+	return enableBackendKeepalive, nil
 }
 
 func getLoadBalancerID(service *v1.Service) string {