+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add support for specifying custom load-balancer names (@grzesiek)
 * Support specifying a fake region by environment variable (@timoreimann)
 * Bump godo dependency to v1.35.1 (@anitgandhi)
+* Add support for enabling backend keepalive feature for load balancers (@anitgandhi)
 
 ## v0.1.23 (beta) - Jan 31th 2020